Added Python and Ruby package modules
Added a module each for Python and Ruby package verification. Each
module exposes an `is_installed` and `version` property. They will each
accept an argument for the path to the preferred package manager binary
to use for determining which packages are installed.

Added integration tests

Updated Dockerfile for Debian Jessie and wrote integration tests to
verify functionality of PythonPackage and RubyPackage modules.